+++ datahub_common/lib/src/api/dto/encoding.dart
@@ -125,6 +125,7 @@ T? decodeEnumNullable<T>(dynamic raw, List<T> values) {
 
 String encodeJsonString(dynamic value) => JsonEncoder().convert(value);
 
+//TODO replace with value.name when upgrading to dart 2.17
 String encodeEnum<T>(T value) => enumName<T>(value);
 
 String? encodeEnumNullable<T>(T? value) {