+++ lib/src/persistence/postgresql/sql/remove_table_builder.dart
@@ -1,5 +1,4 @@
 import 'package:boost/boost.dart';
-import 'package:cl_datahub/cl_datahub.dart';
 
 import 'sql_builder.dart';